requirements.txt
aiodns==3.0.0
aiofile==3.7.2
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
caio==0.9.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
charset-normalizer==2.0.9
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Django==3.2.9
django-cors-headers==3.10.1
djangorestframework==3.12.4
frozenlist==1.2.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
multidict==5.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycparser==2.21
pycryptodomex==3.12.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.7
whitenoise==6.0.0
yarl==1.7.2
zope.interface==5.4.0
